feature/GE-708: moved module dir
