+++ swagger_server/controllers/access_control_controller.py
@@ -24,7 +24,7 @@ from swagger_server.models.user_domain_role import UserDomainRole  # noqa: E501
 from swagger_server.models.user_site_role import UserSiteRole  # noqa: E501
 from swagger_server import util
 
-from core_access_control import db_actions
+from access_control import db_actions
 
 def access_control_roleresourcepermission_delete(role_id, resource_id, permission_id):  # noqa: E501
     """access_control_roleresourcepermission_delete
